+++ CHANGELOG.md
@@ -40,6 +40,7 @@ count (
 - server: refresh JWKs a maximum of once per second (fix #5781)
 - console: action/event trigger transforms are now called REST connectors
 - console: fix list of tables (and schemas) being unsorted when creating a new trigger event (fix #6391)
+- console: fix custom field names breaking browse table sorting and the pre-populating of the edit row form
 - cli: migrate and seed subcommands has an option in prompt to choose and apply operation on all available databases
 - cli: fix `metadata diff --type json | unified-json` behaving incorrectly and showing diff in YAML format.
 - cli: fix regression in `migrate create` command (#7971)